Revert "Don't remove attributes and styles which serialization is omitted"
This reverts commit 8791cc3d9441cdefdd1486966b688b1baeee23a4.
